LibraryUI/src/app/Books/book.service.ts
import { Injectable } from '@angular/core';
import { HttpClient, HttpErrorResponse } from '@angular/common/http';
import { Observable, throwError } from 'rxjs';
import { catchError, tap, map } from 'rxjs/operators';
import { environment } from '../../environments/environment';
import { cBook } from './book-Author';

@Injectable({
  providedIn: 'root'
})
export class bookService {
  private bookUrl = environment.apiEndpoint;

  constructor(private http: HttpClient) { }

  getbooks(): Observable<cBook[]> {
    return this.http.get<cBook[]>(this.bookUrl,{withCredentials:false})
      .pipe(
        tap(data => console.log('All: ' + JSON.stringify(data))),
        catchError(this.handleError)
      );
  }

  
  private handleError(err: HttpErrorResponse): Observable<never> {
    let errorMessage = '';
    if (err.error instanceof ErrorEvent) {
      errorMessage = `An error occurred: ${err.error.message}`;
    } else {
      errorMessage = `Server returned code: ${err.status}, error message is: ${err.message}`;
    }
    console.error(errorMessage);
    return throwError(errorMessage);
  }
}
